+++ settings/common.py
@@ -448,12 +448,15 @@ SOUTH_MIGRATION_MODULES = {
 DEFAULT_AVATAR_SIZE = 80                # 80x80 pixels
 DEFAULT_BIG_AVATAR_SIZE = 300           # 300x300 pixels
 DEFAULT_TIMELINE_IMAGE_SIZE = 640       # 640x??? pixels
+DEFAUL_CARD_IMAGE_WIDTH = 300           # 300 pixels
+DEFAUL_CARD_IMAGE_HEIGHT = 200           # 200 pixels
 
 THUMBNAIL_ALIASES = {
     '': {
         'avatar': {'size': (DEFAULT_AVATAR_SIZE, DEFAULT_AVATAR_SIZE), 'crop': True},
         'big-avatar': {'size': (DEFAULT_BIG_AVATAR_SIZE, DEFAULT_BIG_AVATAR_SIZE), 'crop': True},
         'timeline-image': {'size': (DEFAULT_TIMELINE_IMAGE_SIZE, 0), 'crop': True},
+        'card-image': {'size': (DEFAUL_CARD_IMAGE_WIDTH, DEFAUL_CARD_IMAGE_HEIGHT), 'crop': True},
     },
 }