+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog #
 
 
-## 1.10.0 ????? (unreleased)
+## 1.9.1 Taiga Tribe (unreleased)
 
 ### Features
 - [CSV Reports] Add fields "created_date", "modified_date", "finished_date" to issues CSV report.